tylor_ai/pyproject.toml
[project]
name = "tylor-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "asif-ml-developer", email = "asif.ml.developer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.2",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
tylor-ai = "tylor_ai:main"
simple_agent = "tylor_ai.simple_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
